+++ crates/forge/src/pretty_printing.rs
@@ -82,7 +82,14 @@ pub(crate) fn print_test_result(test_result: &TestCaseSummary) {
         }
     };
 
-    println!("{result_header} {result_name}{fuzzer_report}{result_message}");
+    let block_number_message = match test_result.latest_block_number() {
+        None => String::new(),
+        Some(latest_block_number) => {
+            format!("\nNumber of the block used for fork testing = {latest_block_number}")
+        }
+    };
+
+    println!("{result_header} {result_name}{fuzzer_report}{block_number_message}{result_message}");
 }
 
 pub fn print_failures(all_failed_tests: &[TestCaseSummary]) {