+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - `--color` option to control when colored output is used
+- when specifying `BlockId::Tag(Latest)` block number of the used block will be printed
 
 ## [0.9.1] - 2023-10-30